requirements.txt
kafka-python==2.0.2
confluent-kafka==2.5.3
requests==2.32.3
PyYAML==6.0.2

black